refactor(ldf): extract annotation schemas into luxonis_ml.ldf
Move the LDF annotation models and the ParquetRecord row schema into a
new lightweight `luxonis_ml.ldf` package with its own extra, so consumers
can depend on the data format without pulling in the full data stack.
`luxonis_ml.data.datasets.annotation` and `luxonis_ml.data.utils.parquet`
re-export the moved names, so existing imports keep working.

Pure move: no behavior change.
