+++ luxonis_ml/ldf/requirements.txt
@@ -1,0 +1,4 @@
+numpy~=2.2
+opencv-python~=4.10
+pillow>=10.0
+pycocotools~=2.0
