+++ tests/test_ldf/test_requirements.py
@@ -1,0 +1,12 @@
+"""Packaging coverage for the lightweight LDF extra."""
+
+from pathlib import Path
+
+
+def test_ldf_extra_declares_segmentation_decoders() -> None:
+    requirements = (
+        Path(__file__).parents[2] / "luxonis_ml" / "ldf" / "requirements.txt"
+    ).read_text()
+
+    assert "opencv-python" in requirements
+    assert "pillow" in requirements
